Extracting login form into seperate struct
