+++ login/login_test.go
@@ -8,7 +8,6 @@
 package login
 
 import (
-	"errors"
 	"io/fs"
 	"log"
 	"net/http"
@@ -20,7 +19,6 @@ import (
 	"time"
 
 	oauth2 "github.com/oxisto/oauth2go"
-	"github.com/oxisto/oauth2go/internal/mock"
 	"golang.org/x/crypto/bcrypt"
 )
 
@@ -337,112 +335,6 @@ func Test_handler_ServeHTTP(t *testing.T) {
 			gotCode := rr.Code
 			if tt.wantCode != gotCode {
 				t.Errorf("handler.doLoginPost() header = %v, wantHeader %v", gotCode, tt.wantCode)
-			}
-		})
-	}
-}
-
-func Test_handler_handleLoginPage(t *testing.T) {
-	type fields struct {
-		sessions map[string]*session
-		users    []*User
-		log      oauth2.Logger
-		baseURL  string
-		pwh      PasswordHasher
-		files    fs.FS
-	}
-	type args struct {
-		w         http.ResponseWriter
-		r         *http.Request
-		error     string
-		returnURL string
-	}
-	tests := []struct {
-		name     string
-		fields   fields
-		args     args
-		wantCode int
-		wantBody string
-	}{
-		{
-			name: "invalid fs",
-			args: args{
-				w: httptest.NewRecorder(),
-			},
-			fields: fields{
-				files: &mockFS{OpenError: errors.New("some error")},
-			},
-			wantCode: http.StatusInternalServerError,
-			wantBody: "template: pattern matches no files: `login.html`",
-		},
-		{
-			name: "invalid template",
-			args: args{
-				w: httptest.NewRecorder(),
-			},
-			fields: fields{
-				files: &mockFS{File: &mockFile{content: "{{"}}, // unclosed action
-			},
-			wantCode: http.StatusInternalServerError,
-			wantBody: "template: login.html:1: unclosed action",
-		},
-		{
-			name: "valid template without errors",
-			args: args{
-				w: &mock.MockResponseRecorder{
-					ResponseRecorder: httptest.NewRecorder(),
-				},
-			},
-			fields: fields{
-				files: &mockFS{File: &mockFile{content: "test"}},
-			},
-			wantCode: http.StatusOK,
-			wantBody: "test",
-		},
-		{
-			name: "valid template with error while writing",
-			args: args{
-				w: &mock.MockResponseRecorder{
-					ResponseRecorder: httptest.NewRecorder(),
-					WriteError:       errors.New("some error"),
-				},
-			},
-			fields: fields{
-				files: &mockFS{File: &mockFile{content: "test"}},
-			},
-			wantCode: http.StatusInternalServerError,
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			h := &handler{
-				sessions: tt.fields.sessions,
-				users:    tt.fields.users,
-				log:      tt.fields.log,
-				baseURL:  tt.fields.baseURL,
-				pwh:      tt.fields.pwh,
-				files:    tt.fields.files,
-			}
-
-			h.handleLoginPage(tt.args.w, tt.args.r, tt.args.error, tt.args.returnURL)
-
-			var rr *httptest.ResponseRecorder
-			switch v := tt.args.w.(type) {
-			case *httptest.ResponseRecorder:
-				rr = v
-			case *mock.MockResponseRecorder:
-				rr = v.ResponseRecorder
-			}
-
-			gotCode := rr.Code
-			if tt.wantCode != gotCode {
-				t.Errorf("handler.handleLoginPage() header = %v, wantHeader %v", gotCode, tt.wantCode)
-			}
-
-			gotBody := rr.Body.String()
-			if tt.wantBody != "" && !strings.Contains(gotBody, tt.wantBody) {
-				t.Errorf("handler.handleLoginPage() body = %v, wantBody %v", gotBody, tt.wantBody)
 			}
 		})
 	}
